Move quote generator to services packaged.
