+++ server/src/main/java/at/codecool/humanoraiserver/controller/QuoteController.java
@@ -1,6 +1,5 @@
 package at.codecool.humanoraiserver.controller;
 
-import at.codecool.humanoraiserver.quotegenerator.QuoteGenerator;
 import at.codecool.humanoraiserver.repositories.UsersRepository;
 import at.codecool.humanoraiserver.services.QuotesService;
 import org.springframework.security.core.Authentication;
