+++ server/src/main/java/at/codecool/humanoraiserver/services/QuoteGenerator.java
@@ -1,4 +1,4 @@
-package at.codecool.humanoraiserver.quotegenerator;
+package at.codecool.humanoraiserver.services;
 
 public interface QuoteGenerator {
     String generateQuote() throws InterruptedException;
